api1/requirements.txt
pandas
pandas_datareader
yfinance
matplotlib
unicorn
fastapi_keycloak
httpx
mysql-connector